creating big error util file per francescos comment in PR
