--- packages/instant/src/util/error_description.ts
@@ -1,23 +1,0 @@
-import { AssetBuyerError } from '@0xproject/asset-buyer';
-
-import { assetDataUtil } from '../util/asset_data';
-
-const humanReadableMessageForError = (error: Error, assetData?: string): string | undefined => {
-    if (error.message === AssetBuyerError.InsufficientAssetLiquidity) {
-        const assetName = assetDataUtil.bestNameForAsset(assetData, 'of this asset');
-        return `Not enough ${assetName} available`;
-    }
-
-    return undefined;
-};
-
-export const errorDescription = (error?: any, assetData?: string): { icon: string; message: string } => {
-    let bestMessage: string | undefined;
-    if (error instanceof Error) {
-        bestMessage = humanReadableMessageForError(error, assetData);
-    }
-    return {
-        icon: '😢',
-        message: bestMessage || 'Something went wrong...',
-    };
-};
